избавилась от множества запросов
