+++ main.js
@@ -1,4 +1,5 @@
 import './src/assets/css/style.css';
 import './src/assets/api.js';
+import "./src/init.js"
 import './src/display.js';
 import './src/filter.js';
